move widgetReg to Grid/
